requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
defusedxml==0.4.1
Django==1.8.11
django-analytical==2.2.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-filter==0.13.0
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
django-rest-swagger==0.3.8
kombu==3.0.37
MySQL-python==1.2.5
Pillow==2.8.1
pycountry==1.20
pytz==2016.4
PyYAML==3.11
redis==2.10.6
sqlparse==0.1.19
virtualenv==1.11.6
